+++ search_query/ebsco/v1_0_0/__init__.py
@@ -1,0 +1,13 @@
+"""EBSCO v1.0.0 stubs."""
+
+from .parser import EBSCOParser_v1_0_0, EBSCOListParser_v1_0_0
+from .serializer import EBCOSerializer_v1_0_0
+from .translator import EBCOTranslator_v1_0_0
+
+__all__ = [
+    "EBSCOParser_v1_0_0",
+    "EBSCOListParser_v1_0_0",
+    "EBCOSerializer_v1_0_0",
+    "EBCOTranslator_v1_0_0",
+]
+